UI - OutOfBoardException Exception class is moved into the exceptions subfolder.
