+++ src/main/java/sdmExam/UI/BoardStoneValue.java
@@ -6,19 +6,17 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class BoardStoneValue {
-        private static final Map<Stone, String> StoneValue = new HashMap<>() {
-            {
-                put(Stone.BLACK, "[B]");
-                put(Stone.WHITE, "[W]");
-                put(Stone.NONE, "[ ]");
-            }
-        };
-
-        private BoardStoneValue(){};
-
-        public static String getStoneValue(Stone stone) {
-            return StoneValue.get(stone);
+    private static final Map<Stone, String> StoneValue = new HashMap<>() {
+        {
+            put(Stone.BLACK, "[B]");
+            put(Stone.WHITE, "[W]");
+            put(Stone.NONE, "[ ]");
         }
+    };
 
+    private BoardStoneValue(){}
 
+    public static String getStoneValue(Stone stone) {
+        return StoneValue.get(stone);
+    }
 }
