+++ src/main/java/sdmExam/exceptions/OutOfBoardException.java
@@ -1,4 +1,4 @@
-package sdmExam.UI.Exception;
+package sdmExam.exceptions;
 
 public class OutOfBoardException  extends  Exception{
     public OutOfBoardException(String error)
